Harden prototype consensus and project hygiene

+++ crypto.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include <string>
 #include <vector>
+#include <cstdint>
 #include "block.hpp"
 
 using std::string;
@@ -10,8 +11,9 @@ using std::vector;
 string signMessage(const string& messageHex, const string& privateKeyHex);
 bool verifySignature(const string& messageHex, const string& signatureHex, const string& publicKeyHex);
 string sha256(const string& data);
-double getBalance(const string& pubkey, const vector<Block>& chain);
-double getEffectiveBalanceDuringMining(const string& pubKeyHex,const vector<Block>& chain,const vector<Transaction>& includedTxs);
+int64_t getBalance(const string& pubkey, const vector<Block>& chain);
+int64_t getEffectiveBalanceDuringMining(const string& pubKeyHex, const vector<Block>& chain,
+                                        const vector<Transaction>& includedTxs);
